Finished weighting
The car now moves according to weighting and fastest path algorithm. It also highlights what the fastest path should be based on what the car can see. I made a slight change to the ordering of NORTH, WEST, SOUTH, EAST to the Heading enum in solver.h for debugging purposes, but it shouldn't make a difference to any code as long as you use the Heading enum correctly